requirements.txt
# Core dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
numpy
pandas
propcache==0.4.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.2.0
scipy
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yarl==1.22.0

# Real trader dependencies
py-clob-client>=0.34.0
python-dotenv>=1.0.0
web3>=6.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Database dependencies
psycopg2-binary>=2.9.0
